Workflows: add build and push to Dockerhub workflow
Since now we are using the new docker build infra, the old image
building and publishing workflow, based on dockerhub runners, is broken.
Replace it with a new one, this time fully based on Github to standardize
the image publishing process.
`DOCKER_USERNAME` and `DOCKER_PASSWORD` of the registry must be added
to repo secrets for this to work.
In order to separate the flow of tagged and untagged builds, workflows
are splitted: one works only on tags, the other on everyday push events.

Sorry, dockerhub CI. New kid in town.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>
Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>