DUP on Mississippi: rework message processing infrastructure
Mississippi handles the ordering and delivery of messages,
so all old related infrastructure is to be removed.
Contextually, a number of small changes is made to fit
the framework to our use case (e.g. adding `Mississippi.Consumer`
to the supervision tree, changing the name the
events producer pool so that it does not conflict with Mississippi's).
Contextually, the supervision tree is simplified.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>
Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>