refactor(appengine): exandra queries for device
The first installment in the effort of refactoring the device module.

The change can be made little by little as exandra and cqex queries
can coexist.

This change ports a few queries to exandra, without making any big logic
refactor, but we can see the future direction of this refactor:
Astarte.AppEngine.API.Device.Queries becomes a module to expose _queries_
and not data from the database directly.

For this iteration, this is enough to elegantly use the queries in the
Device module, but more complex queries may require a more in-depth refactor.

The following queries have been ported:
- `retrieve_interfaces_list`
- `retrieve_all_endpoints_for_interface!`
- `retrieve_mapping`
- `prepare_value_type_query` / `execute_value_type_query`

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>